create accounting module
