+++ server/accounting/build.gradle
@@ -1,0 +1,10 @@
+dependencyManagement {
+    imports { mavenBom("org.springframework.boot:spring-boot-dependencies:${version_spring_boot}") }
+}
+
+dependencies {
+    compile(
+            project(":server:user-management"),
+            project(":server:persistence"),
+    )
+}
