+++ server/accounting/src/main/java/de/budgetfreak/accounting/AccountingProperties.java
@@ -1,0 +1,11 @@
+package de.budgetfreak.accounting;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+/**
+ * Properties for the accounting-module.
+ */
+@ConfigurationProperties("accounting")
+public class AccountingProperties {
+
+}
