+++ server/accounting/src/main/java/de/budgetfreak/accounting/web/AccountResource.java
@@ -1,0 +1,28 @@
+package de.budgetfreak.accounting.web;
+
+import org.springframework.hateoas.ResourceSupport;
+
+/**
+ * Ressource representing an account.
+ */
+public class AccountResource extends ResourceSupport {
+
+    private String description;
+    private Boolean onBudget;
+
+    public AccountResource() {
+    }
+
+    public AccountResource(String description, Boolean onBudget) {
+        this.description = description;
+        this.onBudget = onBudget;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public Boolean getOnBudget() {
+        return onBudget;
+    }
+}
