+++ server/accounting/src/test/java/de/budgetfreak/accounting/AccountTestUtils.java
@@ -1,0 +1,31 @@
+package de.budgetfreak.accounting;
+
+import de.budgetfreak.accounting.domain.Account;
+import de.budgetfreak.usermanagement.domain.User;
+
+import java.util.List;
+
+import static java.util.Arrays.asList;
+
+public class AccountTestUtils {
+
+    private AccountTestUtils() {
+    }
+
+    public static List<Account> createAccounts(User user) {
+        return asList(createCheckingsAccount(user), createSavingsAccount(user));
+    }
+
+    public static Account createCheckingsAccount(User user) {
+        return createAccount(1L, "Checkings", true, user);
+    }
+
+    public static Account createSavingsAccount(User user) {
+        return createAccount(2L, "Savings", false, user);
+    }
+
+    public static Account createAccount(long id, String description, boolean onBudget, User user) {
+        Account account = new Account(description, onBudget, user).setId(id);
+        return account;
+    }
+}
