+++ server/accounting/src/test/java/de/budgetfreak/accounting/AccountingTestApplication.java
@@ -1,0 +1,7 @@
+package de.budgetfreak.accounting;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AccountingTestApplication {
+}
