+++ server/user-management/src/test/java/de/budgetfreak/usermanagement/UserManagementTestApplication.java
@@ -1,0 +1,7 @@
+package de.budgetfreak.usermanagement;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UserManagementTestApplication {
+}
