+++ server/user-management/src/test/resources/de/budgetfreak/usermanagement/domain/users.sql
